==> scripts/03_R_analysis/01_filtering_and_normalization.Rmd <==
---
title: "Psilocybin Project 2"
subtitle: "Filtering and Normalization"
author: "Kennedi Todd"
date: "06/03/2024"
output:
  html_document:
    theme: cerulean
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: inline
---

# Background
- bulk RNA sequencing of hemibrain (olfactory bulb and cerebellum removed) of 216 mice \
- 12 samples per group (treatment, dose, time point) \
- within each group there 6 male and 6 female mice \
- one mouse died and one sample had a bad extraction so there's only data on 214 mice \
- high dose psilocybin = 1 mg/kg, low dose psilocybin = 0.25 mg/kg, ketanserin = 2 mg/kg \

# Setup
## Working directory
```{r setup}
knitr::opts_knit$set(root.dir = ".")
```

## Libraries
```{r libraries, message=FALSE, warning=FALSE}
library(caret)              # createDataPartition()
library(dplyr)              # left_join()
library(edgeR)              # DGEList()
library(limma)              # plotMDS()
library(ggrepel)            # geom_text_repel()
library(ggplot2)            # ggplot()
library(glmnet)             # cv.glmnet()
library(gplots)             # heatmap.2()
library(grDevices)          # colorRampPalette()
library(gridExtra)          # grid.arrange()
library(philentropy)        # JSD()
library(rtracklayer)        # import()
library(stringr)            # str_match()
library(tidyverse)          # pivot_wider()
```

## User defined variables
```{r set_variables}
out <- "../../results/all_samples_strict/"
out2 <- "all_samples_strict"
sex_order <- c("Male","Female")
sex_colors <- c("gold2","purple")
group_order <- c("S.8h","K.8h","L.8h","H.8h","KL.8h","KH.8h",
                 "S.24h","K.24h","L.24h","H.24h","KL.24h","KH.24h",
                 "S.7d","K.7d","L.7d","H.7d","KL.7d","KH.7d")
group_colors <- c("#E9E9E9","#FFD0D0","#F7FCB9","#9DF097","#B5DAFF","#CE9FF7",
                  "#9A9A9A","#FA8A8A","#EFF964","#59C751","#60ABF6","#9D54DE",
                  "#545252","#F64646","#CEDC04","#219218","#086CD0","#6612B0")
ann <- paste0("/research/labs/neurology/fryer/projects/references/",
              "mouse/refdata-gex-mm10-2020-A/genes/genes.gtf")
```

## Save functions
These functions with help simultaneously save plots as a png, pdf, and tiff 
file.
```{r save_functions}
saveToPDF <- function(...) {
    d = dev.copy(pdf,...)
    dev.off(d)
}
```

# DGE object
## Read data
```{r read_data}
# read in counts data
counts <- read.delim(paste0("../../featureCounts/gene_counts_matrix.tsv"),
                     header = TRUE,
                     row.names = 1,
                     sep = "\t")

# read metadata
meta <- read.delim("../../refs/metadata.tsv",
                   header = TRUE,
                   sep = "\t")
meta$group <- factor(meta$group, levels = group_order)
meta <- meta %>% ungroup() %>% arrange(group, animal_id)
rownames(meta) <- 1:nrow(meta)

# rename counts columns
counts.names <- gsub("_L00[1-9]", "", colnames(counts))
colnames(counts) <- counts.names

# make count cols match meta rows
counts <- counts[,meta$filename]
all.equal(colnames(counts), meta$filename)

# set row and col names to sample_id
rownames(meta) <- meta$sample_name
colnames(counts) <- meta$sample_name

# read in annotation file
if (file.exists("../../rObjects/annotation.rds")) {
  genes <- readRDS("../../rObjects/annotation.rds")
} else {
  genes <- rtracklayer::import(ann)
  genes <- as.data.frame(genes)
  genes <- genes[genes$type == "gene",]
  all.equal(rownames(counts), genes$gene_id) # check that rows match
  saveRDS(genes, "../../rObjects/annotation.rds")
}

# read exonic length information
exonic.length <- read.delim("../../featureCounts/summed_exons.tsv",
                            header = TRUE,
                            sep = "\t")

# add exonic length to genes
all.equal(rownames(counts), genes$gene_id, exonic.length$gene_id)
genes <- left_join(genes, exonic.length, by = "gene_id")
remove(exonic.length)

# rownames - gene_id to gene_name
all.equal(rownames(counts), genes$gene_id)
rownames(counts) <- make.unique(genes$gene_name)
genes$gene_name_unique <- rownames(counts)
```

## Create object
```{r DGE_object}
# create object
dge <- DGEList(counts = counts,
               genes = genes,
               samples = meta)

# cleanup 
remove(counts, meta, genes)

# set levels
dge$samples$sex <- factor(dge$samples$sex, levels = sex_order)

# run_flowcell_lane
dge$samples$run_flowcell_lane <- gsub("87_22HVM7LT3_8","Run1_Lane8",dge$samples$run_flowcell_lane)
dge$samples$run_flowcell_lane <- gsub("89_22JGCYLT3_5","Run2_Lane5",dge$samples$run_flowcell_lane)
dge$samples$run_flowcell_lane <- gsub("89_22JGCYLT3_6","Run2_Lane6",dge$samples$run_flowcell_lane)
dge$samples$run_flowcell_lane <- gsub("89_22JGCYLT3_7","Run2_Lane7",dge$samples$run_flowcell_lane)
dge$samples$run_flowcell_lane <- gsub("89_22JGCYLT3_8","Run2_Lane8",dge$samples$run_flowcell_lane)
table(dge$samples$run_flowcell_lane)
```

## Remove MT genes
```{r remove_MT_genes}
dim(dge)
removeMT <- dge$genes$seqnames != "chrM"  # true when NOT MT
dge <- dge[removeMT,,keep.lib.sizes = FALSE]
dim(dge)
```

## Keep protein coding genes
```{r remove_MT_genes}
dim(dge)
keep <- dge$genes$gene_type == "protein_coding"
dge <- dge[keep,,keep.lib.sizes = FALSE]
dim(dge)
```

## Optional sample filtering
```{r}
if (out == "../../results/low_lib_removed_strict/") {
  keep <- dge$samples$lib.size > 15000000
  dge <- dge[,keep]
} else if (out == "../../results/8_hours_low_lib_removed_strict/") {
  keep <- dge$samples$timepoint == "8 hour"
  dge <- dge[,keep]
  keep <- dge$samples$lib.size > 15000000
  dge <- dge[,keep]
  group_colors <- c("#9A9A9A","#FA8A8A","#EFF964",
                    "#59C751","#60ABF6","#9D54DE")
} else if (out == "../../results/24_hours_low_lib_removed_strict/") {
  keep <- dge$samples$timepoint == "24 hour"
  dge <- dge[,keep]
  keep <- dge$samples$lib.size > 15000000
  dge <- dge[,keep]
  group_colors <- c("#9A9A9A","#FA8A8A","#EFF964",
                    "#59C751","#60ABF6","#9D54DE")
} else if (out == "../../results/7_days_low_lib_removed_strict/") {
  keep <- dge$samples$timepoint == "7 day"
  dge <- dge[,keep]
  keep <- dge$samples$lib.size > 15000000
  dge <- dge[,keep]
  group_colors <- c("#9A9A9A","#FA8A8A","#EFF964",
                    "#59C751","#60ABF6","#9D54DE")
} else if (out == "../../results/low_lib_removed_lax/") {
  keep <- dge$samples$lib.size > 15000000
  dge <- dge[,keep]
}
dim(dge)
```

# QC: Raw
## Bargraph
```{r cpm_bar_graph}
# user input
goi <- c("Uty","Ddx3y","Kdm5d","Hbb-bs")

# plot and save
for (i in 1:length(goi)) {

  # extract data
  df <- as.data.frame(cpm(dge$counts))
  df <- df[goi[i],]
  df <- reshape2::melt(df)
  colnames(df) <- c("sample","CPM")
  df$sample <- as.character(df$sample)
  df$sex <- dge$samples$sex
  df <- df[order(df$sex),]
  
  # plot
  p <- ggplot(data = df, mapping = aes(x = sample, y = CPM, fill = sex)) +
    geom_col() +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) +
    labs(title = paste("Raw:", goi[i], "CPM"), x = "Sample", y = "CPM") +
    scale_x_discrete(limits = df$sample)
  print(p)
  
  # save
  path <- paste0(out,"prefiltering_QC/", goi[i], "_bar_graph_raw")
  saveToPDF(paste0(path, ".pdf"), width = 25, height = 6)
  dev.off()
}
remove(df,p)
```

## JSD heatmap
This portion won't display in the R Markdown pdf; the margins are too large.
The pdf and png file can only be saved one at a time.
```{r JSD_heatmap}
# find JSD
data <- JSD(t(cpm(dge$counts)), est.prob = "empirical")
names <- dge$samples$sample_name
colnames(data) <- names
rownames(data) <- names
round.data <- round(data, digits = 4) # round 2 decimal places

# save
path <- paste0(out,"prefiltering_QC/JSD_heatmap.pdf")
pdf(path, width = 25, height = 25)

# set heatmap colors and names
meta <- dge$samples[,c("treatment","sex")]
rownames(meta) <- names
paletteLength <- 100
myColor <- colorRampPalette(c("blue", "skyblue", "white"))(paletteLength)
ann_colors = list(sex = c(Male="gold", Female="purple"),
                  treatment = c(Saline="lightgray", 
                                Ketanserin="cornflowerblue",
                                Psilocybin="springgreen2",
                                `Ketanserin+Psilocybin`="tomato"))

pheatmap::pheatmap(round.data,
                   treeheight_row = 0,
                   treeheight_col = 0,
                   annotation_row = meta,
                   annotation_col = meta,
                   annotation_colors = ann_colors,
                   color = myColor)
```

## MDS
- MDS = multidimensional scaling \
- limma::plotMDS() will produce a PCoA or PCA plot \
- By default, gene.selection = "pairwise" \
- If gene.selection = "pairwise" and top is < nrow(x) then a PCoA plot is produced \
- Distances on the plot represent the leading log2 fold-changes \
- The leading log fold-change between a pair of samples is the root-mean-square average of the top largest log2 fold-changes between those two samples \
### Sex
```{r MDS_colored_by_sex}
# set colors and get data
data <- cpm(dge$counts, log = TRUE)
names <- dge$samples$sample_name

# loop through dimensions and top genes
dims <- list(c(1,2),c(2,3),c(3,4))
topGenes <- c(100,500)

for (dim in dims) {
  
  for (numGenes in topGenes) {
      
    mds <- plotMDS(data, 
                   labels = names, 
                   top = numGenes,
                   plot = FALSE,
                   dim.plot = dim)

    toplot <- data.frame(DimX = mds$x, 
                         DimY = mds$y,
                         Sample = rownames(mds$distance.matrix.squared),
                         sex = dge$samples$sex)
    
    x.lab <- paste0(mds$axislabel," ",mds$dim.plot[1]," (",
                    round(mds$var.explained[mds$dim.plot[1]] * 100, digits = 2),
                    "%)")
    y.lab <- paste0(mds$axislabel," ",mds$dim.plot[2]," (",
                    round(mds$var.explained[mds$dim.plot[2]] * 100, digits = 2),
                    "%)")
    
    p <- ggplot(toplot, aes(x = DimX, y = DimY, color = sex, label = Sample)) + 
      geom_point(size = 3) +
      scale_color_manual(values = sex_colors) +
      theme_bw() +
      geom_text(aes(label = Sample), size=4, vjust = -1) +
      xlab(x.lab) + ylab(y.lab) + 
      ggtitle(paste0("MDS Raw: Top ", numGenes, ", Log2(CPM)"))
    print(p)
    
    # save
    dimTitle <- as.character(dim)
    path <- paste0(out,"prefiltering_QC/MDS_top_",
                   numGenes,"_genes/MDS_raw_top_", numGenes, "_",
                   "dim", as.character(dim)[1], "&", as.character(dim[2]), 
                   "_colored_by_sex")
    saveToPDF(paste0(path, ".pdf"), width = 8, height = 8)
  }
}

remove(data)
```

### Group
```{r MDS_colored_by_group}
# set colors and get data
data <- cpm(dge$counts, log = TRUE)
names <- dge$samples$sample_name

# loop through dimensions and top genes
dims <- list(c(1,2),c(2,3),c(3,4),c(4,5))
topGenes <- c(100,500)

for (dim in dims) {
  
  for (numGenes in topGenes) {
      
    mds <- plotMDS(data, 
                   labels = names, 
                   top = numGenes,
                   plot = FALSE,
                   dim.plot = dim)

    toplot <- data.frame(DimX = mds$x, 
                         DimY = mds$y,
                         Sample = rownames(mds$distance.matrix.squared),
                         group = dge$samples$group)
    
    x.lab <- paste0(mds$axislabel," ",mds$dim.plot[1]," (",
                    round(mds$var.explained[mds$dim.plot[1]] * 100, digits = 2),
                    "%)")
    y.lab <- paste0(mds$axislabel," ",mds$dim.plot[2]," (",
                    round(mds$var.explained[mds$dim.plot[2]] * 100, digits = 2),
                    "%)")
    
    p <- ggplot(toplot, aes(x = DimX, y = DimY, color = group, label = Sample)) + 
      geom_point(size = 3) +
      scale_color_manual(values = group_colors) +
      theme_bw() +
      geom_text(aes(label = Sample), size=4, vjust = -1) +
      xlab(x.lab) + ylab(y.lab) + 
      ggtitle(paste0("MDS Raw: Top ", numGenes, ", Log2(CPM)"))
    print(p)
    
    # save
    dimTitle <- as.character(dim)
    path <- paste0(out,"prefiltering_QC/MDS_top_",
                   numGenes,"_genes/MDS_raw_top_", numGenes, "_",
                   "dim", as.character(dim)[1], "&", as.character(dim[2]), 
                   "_colored_by_group")
    saveToPDF(paste0(path, ".pdf"), width = 8, height = 8)
  }
}

remove(data)
```

### Treatment
```{r MDS_colored_by_treatment}
# set colors and get data
data <- cpm(dge$counts, log = TRUE)
names <- dge$samples$sample_name

# loop through dimensions and top genes
dims <- list(c(1,2),c(2,3),c(3,4))
topGenes <- c(100,500)

for (dim in dims) {
  
  for (numGenes in topGenes) {
      
    mds <- plotMDS(data, 
                   labels = names, 
                   top = numGenes,
                   plot = FALSE,
                   dim.plot = dim)

    toplot <- data.frame(DimX = mds$x, 
                         DimY = mds$y,
                         Sample = rownames(mds$distance.matrix.squared),
                         treatment = dge$samples$treatment)
    
    x.lab <- paste0(mds$axislabel," ",mds$dim.plot[1]," (",
                    round(mds$var.explained[mds$dim.plot[1]] * 100, digits = 2),
                    "%)")
    y.lab <- paste0(mds$axislabel," ",mds$dim.plot[2]," (",
                    round(mds$var.explained[mds$dim.plot[2]] * 100, digits = 2),
                    "%)")
    
    p <- ggplot(toplot, aes(x = DimX, y = DimY, color = treatment, label = Sample)) + 
      geom_point(size = 3) +
      #scale_color_manual(values = treatment_colors) +
      theme_bw() +
      geom_text(aes(label = Sample), size=4, vjust = -1) +
      xlab(x.lab) + ylab(y.lab) + 
      ggtitle(paste0("MDS Raw: Top ", numGenes, ", Log2(CPM)"))
    print(p)
    
    # save
    dimTitle <- as.character(dim)
    path <- paste0(out,"prefiltering_QC/MDS_top_",
                   numGenes,"_genes/MDS_raw_top_", numGenes, "_",
                   "dim", as.character(dim)[1], "&", as.character(dim[2]), 
                   "_colored_by_treatment")
    saveToPDF(paste0(path, ".pdf"), width = 8, height = 8)
  }
}

remove(data)
```

### Library size
```{r MDS_colored_by_library_size}
# set colors and get data
data <- cpm(dge$counts, log = TRUE)
names <- dge$samples$sample_name

# loop through dimensions and top genes
dims <- list(c(1,2),c(2,3),c(3,4))
topGenes <- c(100,500)

for (dim in dims) {
  
  for (numGenes in topGenes) {
      
    mds <- plotMDS(data, 
                   labels = names, 
                   top = numGenes,
                   plot = FALSE,
                   dim.plot = dim)

    toplot <- data.frame(DimX = mds$x, 
                         DimY = mds$y,
                         Sample = rownames(mds$distance.matrix.squared),
                         treatment = dge$samples$treatment,
                         library_size = dge$samples$lib.size)
    
    x.lab <- paste0(mds$axislabel," ",mds$dim.plot[1]," (",
                    round(mds$var.explained[mds$dim.plot[1]] * 100, digits = 2),
                    "%)")
    y.lab <- paste0(mds$axislabel," ",mds$dim.plot[2]," (",
                    round(mds$var.explained[mds$dim.plot[2]] * 100, digits = 2),
                    "%)")
    
    p <- ggplot(toplot, aes(x = DimX, y = DimY, 
                            color = library_size, label = Sample)) + 
      geom_point(size = 3) +
      scale_color_gradientn(colors = c("red","yellow","lightblue","blue")) +
      theme_bw() +
      geom_text(aes(label = Sample), size=4, vjust = -1) +
      xlab(x.lab) + ylab(y.lab) + 
      ggtitle(paste0("MDS Raw: Top ", numGenes, ", Log2(CPM)"))
    print(p)
    
    # save
    dimTitle <- as.character(dim)
    path <- paste0(out,"prefiltering_QC/MDS_top_",
                   numGenes,"_genes/MDS_raw_top_", numGenes, "_",
                   "dim", as.character(dim)[1], "&", as.character(dim[2]), 
                   "_colored_by_library_size")
    saveToPDF(paste0(path, ".pdf"), width = 8, height = 8)
  }
}

remove(data)
```

### Batch
```{r MDS_colored_by_batch}
# set colors and get data
data <- cpm(dge$counts, log = TRUE)
names <- dge$samples$sample_name

# loop through dimensions and top genes
dims <- list(c(1,2),c(2,3),c(3,4))
topGenes <- c(100,500)

for (dim in dims) {
  
  for (numGenes in topGenes) {
      
    mds <- plotMDS(data, 
                   labels = names, 
                   top = numGenes,
                   plot = FALSE,
                   dim.plot = dim)

    toplot <- data.frame(DimX = mds$x, 
                         DimY = mds$y,
                         Sample = rownames(mds$distance.matrix.squared),
                         treatment = dge$samples$treatment,
                         sequencing_batch = dge$samples$run_flowcell_lane)
    
    x.lab <- paste0(mds$axislabel," ",mds$dim.plot[1]," (",
                    round(mds$var.explained[mds$dim.plot[1]] * 100, digits = 2),
                    "%)")
    y.lab <- paste0(mds$axislabel," ",mds$dim.plot[2]," (",
                    round(mds$var.explained[mds$dim.plot[2]] * 100, digits = 2),
                    "%)")
    
    p <- ggplot(toplot, aes(x = DimX, y = DimY, 
                            color = sequencing_batch, label = Sample)) + 
      geom_point(size = 3) +
      theme_bw() +
      geom_text(aes(label = Sample), size=4, vjust = -1) +
      xlab(x.lab) + ylab(y.lab) + 
      ggtitle(paste0("MDS Raw: Top ", numGenes, ", Log2(CPM)"))
    print(p)
    
    # save
    dimTitle <- as.character(dim)
    path <- paste0(out,"prefiltering_QC/MDS_top_",
                   numGenes,"_genes/MDS_raw_top_", numGenes, "_",
                   "dim", as.character(dim)[1], "&", as.character(dim[2]), 
                   "_colored_by_sequencing_batch")
    saveToPDF(paste0(path, ".pdf"), width = 8, height = 8)
  }
}

remove(data)
```

## CPM, FPKM, TPM
```{r cpm_fpkm_tpm}
# CPM, Counts Per Million
# normalize by sequencing depth
cpm <- cpm(dge)
path <- paste0(out, "counts/CPM_prefiltering.tsv")
write.table(cpm, path, quote = FALSE, sep = "\t")

# FPKM, Fragments Per Kilobase Million
# normalize by sequencing depth and then length
exonic.length <- dge$genes$exonic_length
exonic.length.kb <- exonic.length / 1000
fpkm <- cpm / exonic.length.kb
fpkm <- sweep(cpm, 1, exonic.length.kb, FUN = '/') # 1 = row
path <- paste0(out, "counts/FPKM_prefiltering.tsv")
write.table(fpkm, path, quote = FALSE, sep = "\t")

# TPM, Transcripts Per Kilobase Million
# normalize by length and then sequencing depth
counts <- dge$counts
fpk <- counts / exonic.length.kb
sum.col <- colSums(fpk) / 1000000
tpm <- sweep(fpk, 2, sum.col, FUN = '/') # 2 = column
path <- paste0(out, "counts/TPM_prefiltering.tsv")
write.table(tpm, path, quote = FALSE, sep = "\t")

remove(counts,cpm,fpk,fpkm,tpm)
```

## Library
```{r}
# outlier function
is_outlier <- function(x) {
  return(x < quantile(x, 0.25) - 1.5 * IQR(x) | x > quantile(x, 0.75) + 1.5 * IQR(x))
}

# extract data
df <- data.frame(project = "Psi2_all_samples",
                 sample = dge$samples$sample_name,
                 lib = dge$samples$lib.size,
                 log2.lib = log2(dge$samples$lib.size),
                 group = dge$samples$group)

df %>%
  group_by(project) %>%
  mutate(outlier = ifelse(is_outlier(lib), lib, as.numeric(NA))) %>%
  ggplot(., aes(x = factor(project), y = lib)) +
    geom_boxplot() +
    geom_text(aes(label = sample), na.rm = TRUE, hjust = -0.3)

# plot
b1 <- ggplot(data = df, mapping = aes(x = project, y = lib)) +
  geom_boxplot() +
  geom_text(aes(label = sample), na.rm = TRUE, hjust = -0.3) +
  labs(y = "Library Size", x ="Samples") +
  geom_hline(yintercept = 15000000, color = "red", linetype = "dashed")
b1
# save
path <- paste0(out,"prefiltering_QC/library_boxplot_raw")
saveToPDF(paste0(path, ".pdf"), width = 10, height = 6)
dev.off()

# plot
b2 <- ggplot(data = df, mapping = aes(x = group, y = lib, fill = group)) +
  geom_boxplot() +
  geom_text(aes(label = sample), na.rm = TRUE, hjust = -0.3) +
  labs(y = "Library Size", x ="Group") +
  geom_hline(yintercept = 15000000, color = "red", linetype = "dashed")
b2
# save
path <- paste0(out,"prefiltering_QC/library_split_group_boxplot_raw")
saveToPDF(paste0(path, ".pdf"), width = 18, height = 6)
dev.off()

# plot
b3 <- ggplot(data = df, mapping = aes(x = project, y = log2.lib)) +
  geom_boxplot() +
  geom_text(aes(label = sample), na.rm = TRUE, hjust = -0.3)
b3
# save
path <- paste0(out,"prefiltering_QC/log2_library_boxplot_raw")
saveToPDF(paste0(path, ".pdf"), width = 6, height = 6)
dev.off()
```

# Filtering and normalization
## Lowly expressed genes
```{r filter_genes_by_expr}
start <- dim(dge)[1]

# get dataframes
cpm <- as.data.frame(cpm(dge))
meta <- dge$samples[,c("sample_name","group")]
meta <- meta |> group_by(group) |> mutate(group_size = n())
meta$min_samples <- floor(meta$group_size * 0.8)

# find CPM avg per group
mean.group <- cpm |>
  mutate(gene = row.names(cpm)) |>
  pivot_longer(-gene, names_to = "sample_name") |>
  left_join(meta, by = "sample_name") |>
  summarize(mean = mean(value), .by = c(gene, group)) |>
  pivot_wider(names_from = group, values_from = mean)
mean.group <- as.data.frame(mean.group)
rownames(mean.group) <- mean.group$gene
mean.group$gene <- NULL
mean.group <- round(mean.group, digits = 4)
# save
write.table(x = mean.group, 
            file = paste0(out,"counts/avgCPM_per_group_before_filtering.tsv"),
            quote = FALSE,
            sep = "\t")

# Option 1 filtering method
# floor(80% samples of a group) must have CPM >= 1
strict <- as.data.frame(cpm >= 1) |> 
  mutate(gene = rownames(cpm)) |>
  pivot_longer(-gene, names_to = "sample_name") |>
  left_join(meta, by = "sample_name") |>
  summarize(samples_meeting_thresh = sum(value), 
            .by = c(gene, group, group_size, min_samples))|>
  mutate(meets_thresh = samples_meeting_thresh >= min_samples) |>
  select (-c(group_size, min_samples, samples_meeting_thresh)) |>
  pivot_wider(names_from = group, values_from = meets_thresh)
strict <- as.data.frame(strict)
rownames(strict) <- strict$gene
strict$gene <- NULL
strict$pass <- apply(X = strict, MARGIN = 1, FUN = any)
all.equal(rownames(strict), rownames(dge))
# check with Myocd

# all samples in a group (i.e. sal.8h) must have CPM >= 1
# create df with summary of if group has all samples >= 1
#strict <- as.data.frame(cpm >= 1) |> 
#  mutate(gene = rownames(cpm)) |>
#  pivot_longer(-gene, names_to = "sample_name") |>
#  left_join(meta, by = "sample_name") |>
#  summarize(thresh = all(value), .by = c(gene, group)) |>
#  pivot_wider(names_from = group, values_from = thresh, )
# summarize df - if any group meets the thresh then TRUE
#strict <- as.data.frame(strict)
#rownames(strict) <- strict$gene
#strict$gene <- NULL
#strict$pass <- apply(X = strict, MARGIN = 1, FUN = any)
#all.equal(rownames(strict),rownames(dge))

# filter
dge.filtered <- dge[strict$pass,,keep.lib.sizes = FALSE]

# save
mean.group.final <- mean.group[strict$pass,]
write.table(x = mean.group.final, 
            file = paste0(out, "counts/avgCPM_per_group_after_filtering.tsv"),
            quote = FALSE,
            sep = "\t")

end <- dim(dge.filtered)[1]
print(paste(start - end, "genes removed and", end, "genes leftover."))
remove(meta,mean.group,cpm,start,end,mean.group.final,strict)
```

## TMM
Convert observed library sizes into effective library sizes. To obtain effective
library sizes, the current library sizes are multiplied by a normalization factor
which we find using the trimmed mean of M values (TMM) method. The TMM method
trims observations from each tail of the distribution of log-rations (M-values). 
By default this method uses a sumTrim of 0.05 and a logratioTrim of 0.3.
```{r TMM_normalize}
# Normalize by trimmed mean of M-values
dge.filtered.norm <- calcNormFactors(dge.filtered, method = "TMM")

# Normalization factor summary
summary(dge.filtered.norm$samples$norm.factors)
```

# QC: After normalization
## Density plot
Density plots of log-intensity distribution of each library can be superposed 
on a single graph for a better comparison between libraries and for 
identification of libraries with weird distribution. 
```{r density_plots}
# set graphical parameter
par(mfrow = c(1,3))

# Normalize data for library size and expression intesntiy
log2cpm.raw <- cpm(dge, log = TRUE)
colnames(log2cpm.raw) <- dge$samples$sample_name
log2cpm.filtered <- cpm(dge.filtered, log = TRUE)
colnames(log2cpm.filtered) <- dge.filtered$samples$sample_name
log2cpm.norm <- cpm(dge.filtered.norm, log = TRUE)
colnames(log2cpm.norm) <- dge.filtered.norm$samples$sample_name

# set colors
colors <- group_colors[dge.filtered.norm$samples$group]
nsamples <- ncol(dge)

# First, plot the first column of the log2cpm.raw density
plot(density(log2cpm.raw[,1]), col = colors[1], lwd = 2, ylim = c(0,0.25), 
     las = 2, main = "A. Raw", xlab = expression('Log'[2]~CPM))

# For each sample plot the lcpm density
for (i in 2:nsamples){
  den <- density(log2cpm.raw[,i]) #subset each column
  lines(den$x, den$y, col = colors[i], lwd = 2) 
}

# Second, plot log2cpm.filtered
plot(density(log2cpm.filtered[,1]), col = colors[1], lwd = 2, ylim = c(0,0.25), 
     las = 2, main = "B. Filtered", xlab = expression('Log'[2]~CPM))
abline(v = cpm(3, log = TRUE), lty = 3)
for (i in 2:nsamples) {
  den <- density(log2cpm.filtered[,i])
  lines(den$x, den$y, col = colors[i], lwd = 2)
}

# Third, plot log2cpm.norm
plot(density(log2cpm.norm[,1]), col = colors[1], lwd = 2, ylim = c(0,0.25), 
     las = 2, main = "C. Normalized", xlab = expression('Log'[2]~CPM))
abline(v = cpm(3, log = TRUE), lty = 3)
for (i in 2:nsamples) {
  den <- density(log2cpm.norm[,i])
  lines(den$x, den$y, col = colors[i], lwd = 2)
}

# save
path <- paste0(out, "postfiltering_QC/density_plots")
saveToPDF(paste0(path, ".pdf"), width = 6, height = 4)
remove(den)
```

## Boxplots
```{r boxplots}
# set parameters
par(mfrow = c(3,1))

# First look at dge
boxplot(log2cpm.raw, 
        main="A. Raw", 
        xlab="", 
        ylab=expression('Counts per gene (Log'[2]~'CPM)'),
        axes=FALSE,
        col = colors
        )
axis(2) # 2 = left 
axis(1, # 1 = below 
     at = 1:nsamples, # points at which tick-marks should be drawn
     labels = colnames(log2cpm.raw),
     las = 2,
     cex.axis = 0.8 # size of axis
     )

# Second, look at dge.filtered
boxplot(log2cpm.filtered, 
        main="B. Filtered", 
        xlab="", 
        ylab=expression('Counts per gene (Log'[2]~'CPM)'),
        axes=FALSE,
        col = colors
        )
axis(2)
axis(1, at=1:nsamples,labels=colnames(log2cpm.filtered),las=2,cex.axis=0.8)

# Third, look at dge.norm
boxplot(log2cpm.norm, 
        main="C. Normalized", 
        xlab="", 
        ylab=expression('Counts per gene (Log'[2]~'CPM)'),
        axes=FALSE,
        col = colors)
axis(2)
axis(1,at=1:nsamples,labels=colnames(log2cpm.norm),las=2,cex.axis=0.8)

# save
path <- paste0(out, "postfiltering_QC/counts_per_gene_boxplot")
saveToPDF(paste0(path, ".pdf"), width = 25, height = 20)
#remove(log2cpm.raw, log2cpm.norm, log2cpm.filtered)
```

## CPM, FPKM, TPM
```{r cpm_fpkm_tpm}
# CPM, Counts Per Million
# normalize by sequencing depth
cpm <- cpm(dge.filtered.norm)
path <- paste0(out, "counts/CPM_postfiltering.tsv")
write.table(cpm, path, quote = FALSE, sep = "\t")

# FPKM, Fragments Per Kilobase Million
# normalize by sequencing depth and then length
exonic.length <- dge.filtered.norm$genes$exonic_length
exonic.length.kb <- exonic.length / 1000
fpkm <- cpm / exonic.length.kb
fpkm <- sweep(cpm, 1, exonic.length.kb, FUN = '/') # 1 = row
path <- paste0(out, "counts/FPKM_postfiltering.tsv")
write.table(fpkm, path, quote = FALSE, sep = "\t")

# TPM, Transcripts Per Kilobase Million
# normalize by length and then sequencing depth
counts <- dge.filtered.norm$counts
fpk <- counts / exonic.length.kb
sum.col <- colSums(fpk) / 1000000
tpm <- sweep(fpk, 2, sum.col, FUN = '/') # 2 = column
path <- paste0(out, "counts/TPM_postfiltering.tsv")
write.table(tpm, path, quote = FALSE, sep = "\t")

remove(counts,cpm,fpk,fpkm,tpm)
```

## Top transcripts
```{r}
counts <- as.data.frame(dge.filtered.norm$counts)
df <- as.data.frame(rowSums(counts))
colnames(df) <- "rsum"
df$gene <- rownames(df)
df <- df[order(df$rsum, decreasing = TRUE),]
rownames(df) <- 1:nrow(df)
head(df, 20)

path <- paste0(out, "counts/top_transcripts_counts_postfiltering.tsv")
write.table(df, path, quote = FALSE, sep = "\t")
```

# Save
```{r save_final_object, eval=FALSE}
saveRDS(dge.filtered.norm, 
        paste0("../../rObjects/", out2, "_DGEList_filtered_normalized.rds"))
```

```{r session_info}
sessionInfo()
```
