==> scripts/03_R_analysis/02_variance.Rmd <==
---
title: "Psilocybin Project 2"
subtitle: "Variance"
author: "Kennedi Todd"
date: "05/27/2024"
output:
  html_document:
    theme: cerulean
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: inline
---

# Setup
## Working directory
```{r working_directory}
knitr::opts_knit$set(root.dir = ".")
```

## Libraries
```{r libraries}
library(corrplot)          # corrplot()
library(dplyr)             # %>%
library(edgeR)             # load DGEList object
library(ggcorrplot)        # ggcorrplot()
library(gplots)            # heatmap.2()
library(limma)             # voomWithQualityWeights()
library(mvIC)              # mvForwardStepwise()
library(variancePartition) # fitExtractVarPartModel()
```

## User defined variables
```{r set_variables}
out <- "../../results/7_days_low_lib_removed_strict/"
out2 <- "7_days_low_lib_removed_strict"
sex_order <- c("Male","Female")
sex_colors <- c("gold2","purple")
group_order <- c("S.8h","K.8h","L.8h","H.8h","KL.8h","KH.8h",
                 "S.24h","K.24h","L.24h","H.24h","KL.24h","KH.24h",
                 "S.7d","K.7d","L.7d","H.7d","KL.7d","KH.7d")
group_colors <- c("#E9E9E9","#FFD0D0","#F7FCB9","#9DF097","#B5DAFF","#CE9FF7",
                  "#9A9A9A","#FA8A8A","#EFF964","#59C751","#60ABF6","#9D54DE",
                  "#545252","#F64646","#CEDC04","#219218","#086CD0","#6612B0")
```

## Load data
```{r load_data}
dge.filtered.norm <- readRDS(paste0("../../rObjects/",
                                    out2, 
                                    "_DGEList_filtered_normalized.rds"))
```

# Cell Markers
## Extract metadata
```{r extract_meta}
# get meta
meta <- as.data.frame(dge.filtered.norm$samples)

# select variables of interest
variables <- c("group",
               "treatment",
               "dose",
               "timepoint",
               "lib.size",
               "sex",
               "run_flowcell_lane")
meta <- meta %>% select(all_of(variables))

# log2 transform lib.size
meta$log2.lib <- log2(meta$lib.size)
meta$lib.size <- NULL

# change class
meta$dose <- factor(meta$dose)
meta$timepoint <- factor(meta$timepoint)
meta$run_flowcell_lane <- factor(meta$run_flowcell_lane)
meta$treatment <- factor(meta$treatment)

# check classes
sapply(meta, class)
```

## Markers
- Astrocytes: GFAP, GJA1, CLU, AQP4 \
- Choroid: TTR \
- Endothelial: FLT1, PECAM1, VWF \
- Fibroblasts: COL1A1, COL1A2, LUM, DCN \
- Macrophages/Microglia: MRC1, MKI67, C1QB \
- Neurons: GAD1, GAD2, SNAP25, SYT1, MEG3 \
- Oligodendrocytes: MBP \
- OPCs: OLIG1, TNR \
- Pericytes & SMCs: MYH11, ACTA2, VTN \
- RBCs - HBB, HBA1, HBA2 \
```{r markers}
# choose cell type markers
markers <- c("Gfap","Ttr","Flt1","Col1a1","C1qb","Syt1","Mbp","Olig1",
             "Acta2","Hbb-bs")
length(markers)

# print markers not in annotation (could have alternative name)
markers[!markers %in% rownames(dge.filtered.norm)]

# get counts
cpm <- round(as.data.frame(cpm(dge.filtered.norm$counts, log = TRUE)), digits = 2)

# filter counts by marker
cpm <- cpm[rownames(cpm) %in% markers,]

# switch cols and rows
cpm <- t(cpm)

# add to meta
all.equal(rownames(meta), rownames(cpm))
meta <- cbind(meta, cpm)
```

## Significance
```{r stats}
# Set comparison values
x <- "S.7d" # control
y <- "KL.7d" # treatment

# preview variables - only want numerical values
colnames(meta)

# initialize table
variable_numbers <- c(6:16)
variable_names <- colnames(meta)[variable_numbers]
tests <- c("control_mean","treatment_mean",
           "control_stdv","treatment_stdv",
           "control_shapiro_pval","treatment_shapiro_pval",
           "levene_na_omit",
           "t_test_t","t_test_pval",
           "wilcoxon_pval")
stats <- matrix(0,
                nrow = length(variable_numbers),
                ncol = length(tests),
                dimnames = list(variable_names, tests))
stats <- as.data.frame(stats)

# Loop through variables
for (i in 1:length(variable_names)) {
  
  # Subset values by group
  values <- as.numeric(meta[, variable_names[i]])
  group <- factor(meta$group)
  control_values <- meta[meta$group == x, variable_names[i]]
  treatment_values <- meta[meta$group == y, variable_names[i]]
  
  # Welch two sample t-test
  t_test <- t.test(control_values, treatment_values)
  t_test_t <- t_test$statistic
  t_test_pval <- t_test$p.value
  
  # Add to table
  stats[i, "control_mean"] <- round(as.numeric(t_test$estimate[1]),4)
  stats[i, "treatment_mean"] <- round(as.numeric(t_test$estimate[2]),4)
  stats[i, "t_test_t"] <- round(t_test_t,4)
  stats[i, "t_test_pval"] <- round(t_test_pval,4)
  
  # Shapiro-Wilk test, normality test
  control_shapiro <- shapiro.test(control_values)
  treatment_shapiro <- shapiro.test(treatment_values)
  
  # Add to table
  stats[i, "control_shapiro_pval"] <- round(control_shapiro$p.value,4)
  stats[i, "treatment_shapiro_pval"] <- round(treatment_shapiro$p.value,4)
  
  # Variance and standard deviation
  control_var <- var(control_values)
  treatment_var <- var(treatment_values)
  
  # Add to table 
  stats[i, "control_stdv"] <- round(sqrt(control_var),4)
  stats[i, "treatment_stdv"] <- round(sqrt(treatment_var),4)
  
  # Two-sample Wilcoxon / Mann-Whitney
  wil <- wilcox.test(control_values, treatment_values)
  
  # Add to table
  stats[i, "wilcoxon_pval"] <- round(wil$p.value, 4)
  
  # Levene's test
  levTest <- car::leveneTest(values ~ group)
  lev <- levTest$`Pr(>F)`
  lev_naomit <- as.numeric(na.omit(lev))
  
  # Add to table
  stats[i, "levene_na_omit"] <- round(lev_naomit, 4)
}

new.names <- gsub("control", x, colnames(stats))
new.names <- gsub("treatment", y, new.names)
colnames(stats) <- new.names

write.table(x = stats,
            file = paste0(out, "variance/stats_", x, "_vs_", y, ".tsv"),
            quote = FALSE,
            sep = "\t")

remove(control_shapiro,treatment_shapiro,
       t_test, wil, levTest)
```

# Correlation
## Spearman correlation
- Spearman's correlation can be used to analyze categorical data, as well as continuous data, because it uses ranks instead of assumptions of normality plotCorrMatrix(cor(info)) \
```{r spearman_heatmap}
# save
path <- paste0(out, "variance/spearman_correlation_heatmap1.pdf")
pdf(path, width = 10, height = 10)
# formula
form <- ~0 + treatment + dose + log2.lib + sex + run_flowcell_lane
# correlation plot
model.matrix(form, data = meta) %>% 
  cor(method = "spearman") %>% 
  ggcorrplot(show.diag = FALSE, type = "lower", lab = TRUE, lab_size = 2)
dev.off()

# save
path <- paste0(out, "variance/spearman_correlation_heatmap2.pdf")
pdf(path, width = 10, height = 10)
# formula
form <- ~0 + group + log2.lib + sex + run_flowcell_lane
# correlation plot
model.matrix(form, data = meta) %>% 
  cor(method = "spearman") %>% 
  ggcorrplot(show.diag = FALSE, type = "lower", lab = TRUE, lab_size = 2)
dev.off()
```

## CCA heatmap
```{r CCA_heatmap}
# save
path <- paste0(out, "variance/canonical_correlation_analysis_heatmap1.pdf")
pdf(path, width = 8, height = 8)
# formula
form <- ~0 + treatment + dose + sex + log2.lib + run_flowcell_lane
c <- canCorPairs(form, meta)
corrplot(c,
         type = "lower",
         is.corr = FALSE,
         order = "hclust",
         col = colorRampPalette(c("blue4", "deepskyblue", "white"))(100),
         addCoef.col = TRUE,
         method = "shade",
         tl.col = "black",
         tl.cex = 0.8,
         number.cex = 0.7,
         diag = TRUE)
dev.off()

# save
path <- paste0(out, "variance/canonical_correlation_analysis_heatmap2.pdf")
pdf(path, width = 8, height = 8)
# formula
form <- ~0 + group + sex + log2.lib + run_flowcell_lane
c <- canCorPairs(form, meta)
corrplot(c,
         type = "lower",
         is.corr = FALSE,
         order = "hclust",
         col = colorRampPalette(c("blue4", "deepskyblue", "white"))(100),
         addCoef.col = TRUE,
         method = "shade",
         tl.col = "black",
         tl.cex = 0.8,
         number.cex = 0.7,
         diag = TRUE)
dev.off()
```

# Variance partition
## Variance model
```{r variance_partition, message=FALSE, eval=FALSE}
# variables to choose from
geneExpr <- as.matrix(dge.filtered.norm$counts)
colnames(meta)
# age is usually a continuous so model it as a fixed effect "age"
# group is categorical, so model them as random effects "(1|group)"
form <- ~ (1|treatment) + (1|dose) + (1|sex)
varPart <- fitExtractVarPartModel(geneExpr, form, meta)
vp <- sortCols(varPart)
saveRDS(vp, paste0("../../rObjects/", out2, "_var_par1.rds"))

# variables to choose from
geneExpr <- as.matrix(dge.filtered.norm$counts)
colnames(meta)
# age is usually a continuous so model it as a fixed effect "age"
# group is categorical, so model them as random effects "(1|group)"
form <- ~ (1|group) + (1|sex)
varPart <- fitExtractVarPartModel(geneExpr, form, meta)
vp <- sortCols(varPart)
saveRDS(vp, paste0("../../rObjects/", out2, "_var_par2.rds"))
```

## Violins
```{r variance_violins, eval=FALSE}
# save
path <- paste0(out, "variance/variance_violins2.pdf")
pdf(path, width = 12, height = 6)

# plot
plotVarPart(vp)
```

## Percent bars
```{r variance_percent_bars}
# save
path <- paste0(out, "variance/variance_gene_percent_bars2.pdf")
pdf(paste0(path, ".pdf"), width = 12, height = 6)

# plot
plotPercentBars(vp[1:10,])
```

## Group var
```{r variance_group}
# sort genes based on variance explained by group
varPart.df <- as.data.frame(vp)
order.varPart.df <- varPart.df[order(varPart.df$group, decreasing = TRUE),]
head(order.varPart.df["group"], 10)
```

## Sex var
```{r variance_group}
# sort genes based on variance explained by group
order.varPart.df <- varPart.df[order(varPart.df$sex, decreasing = TRUE),]
head(order.varPart.df["sex"], 10)
```

## Lib size var
```{r variance_group}
# sort genes based on variance explained by group
order.varPart.df <- varPart.df[order(varPart.df$log2.lib, decreasing = TRUE),]
head(order.varPart.df["log2.lib"], 10)
```

# Finding best model
variancePartition::voomWithDreamWeights() is the same as limma::voom(), except that it allows random effects in the formula. limma::voomWithQualityWeights() includes observational and sample specific quality weights whereas voomWithDreamWeights only includes observational quality weights.

## Forward stepwise regression
```{r find_best_model}
# get matrix of expression data
v <- voomWithDreamWeights(counts = dge.filtered.norm$counts,
                          formula = (~ 0 + group + sex),
                          data = dge.filtered.norm$samples,
                          BPPARAM = BiocParallel::SnowParam(cores),
                          plot = FALSE)
counts <- v$E
colnames(meta) <- make.names(colnames(meta))

# set base formula
base <- ~ (1 | group) + (1 | sex)

# variable we wish to test
variables <- c("log2.lib",
               "RIN",
               "C1qb",
               "Flt1",
               #"Hbb.bs",
               "Syt1",
               "Col1a1",
               "Gfap",
               "Olig1",
               "Ttr",
               "Mbp",
               "Acta2")

# multivariate forward stepwise regression evaluated by multivariate BIC
bestModel <- mvForwardStepwise(exprObj = counts,
                               baseFormula = base,
                               data = meta,
                               variables = variables)
```

```{r session_info}
sessionInfo()
```
